Change the preferred node affinity of pods before the required one
When any errors occur in the required node affinity processing, we should not skip the preferred affinities as they are unrelated.

See, the previously broken/failing tests, e.g.:

"When deploying workloads with public and private images", it "should not set the required node affinity if missing a pull secret".